requirements.txt
# Django y REST Framework
django~=4.2
djangorestframework~=3.14
djangorestframework-simplejwt~=5.3

# Configuración y Base de Datos
python-decouple~=3.8
dj-database-url~=2.1
psycopg2-binary~=2.9  # Driver para PostgreSQL (esencial para Render)

# Servidor de producción y archivos estáticos
gunicorn~=22.0
whitenoise~=6.6

# APIs de Google
google-api-python-client~=2.131
google-auth-oauthlib~=1.2
google-auth-httplib2~=0.2
google-cloud-storage~=2.14.0  # Para almacenamiento de imágenes

# Monitoring y Logging
sentry-sdk~=1.40

# Utilidades
requests~=2.31
python-dateutil~=2.8
pytz~=2023.3
django-cors-headers~=4.3
APScheduler==3.10.4
django-ratelimit~=4.1
pillow~=10.2.0  # Para procesamiento de imágenes

